+++ packages/react-ui/src/widgets/FeatureSelectionWidgetUI/FeatureSelectionWidgetUI.js
@@ -1,9 +1,11 @@
 import React, { useMemo } from 'react';
 import { Paper, capitalize, styled } from '@mui/material';
+import { useIntl } from 'react-intl';
 import PropTypes from 'prop-types';
 import FeatureSelectionUIToggleButton from './FeatureSelectionUIToggleButton';
 import FeatureSelectionUIGeometryChips from './FeatureSelectionUIGeometryChips';
 import FeatureSelectionUIDropdown from './FeatureSelectionUIDropdown';
+import useImperativeIntl from '../../hooks/useImperativeIntl';
 
 const StylesWrapper = styled(Paper)(({ theme }) => ({
   display: 'flex',
@@ -39,37 +41,77 @@ const StylesWrapper = styled(Paper)(({ theme }) => ({
  *
  * -->
  */
-function FeatureSelectionWidgetUI({
-  selectionModes,
-  editModes = [],
-  selectedMode,
-  onSelectMode,
-  enabled,
-  onEnabledChange,
-  geometry,
-  onSelectGeometry,
-  onDeleteGeometry,
-  tooltipPlacement = 'bottom',
-  size = 'medium',
-  chipLabel
-}) {
+function FeatureSelectionWidgetUI(props) {
+  const {
+    selectionModes,
+    editModes = [],
+    selectedMode,
+    onSelectMode,
+    enabled,
+    onEnabledChange,
+    geometry,
+    onSelectGeometry,
+    onDeleteGeometry,
+    tooltipPlacement = 'bottom',
+    size = 'medium',
+    chipLabel
+  } = props;
+  const intl = useIntl();
+  const intlConfig = useImperativeIntl(intl);
+
+  const selectionModeWithLocales = useMemo(() => {
+    return selectionModes.map((mode) => ({
+      ...mode,
+      label: intlConfig.formatMessage({
+        id: `c4r.widgets.featureSelection.${getModeKeyForLocale(mode.id)}`
+      })
+    }));
+  }, [selectionModes, intlConfig]);
+
+  const editModesWithLocales = useMemo(() => {
+    return editModes.map((mode) => ({
+      ...mode,
+      label: intlConfig.formatMessage({
+        id: `c4r.widgets.featureSelection.${getModeKeyForLocale(mode.id)}`
+      })
+    }));
+  }, [editModes, intlConfig]);
+
   const selectedModeData = useMemo(() => {
     const modes = [
-      ...selectionModes.map((mode) => ({ ...mode, isEdit: false })),
-      ...editModes.map((mode) => ({ ...mode, isEdit: true }))
+      ...selectionModeWithLocales.map((mode) => ({
+        ...mode,
+        isEdit: false
+      })),
+      ...editModesWithLocales.map((mode) => ({
+        ...mode,
+        isEdit: true
+      }))
     ];
+
     const foundMode = modes.find((m) => m.id === selectedMode);
     if (!foundMode) {
       throw new Error(`Selected mode "${selectedMode}" not supported`);
     }
     return foundMode;
-  }, [editModes, selectionModes, selectedMode]);
+  }, [editModesWithLocales, selectionModeWithLocales, selectedMode]);
 
   const label = capitalize(selectedModeData?.label || '');
-  const hoverTooltip = selectedModeData?.isEdit ? label : `Select a ${label}`;
+  const hoverTooltip = selectedModeData?.isEdit
+    ? label
+    : intlConfig.formatMessage(
+        {
+          id: `c4r.widgets.featureSelection.selectTool`
+        },
+        { label }
+      );
   const clickTooltip = selectedModeData?.isEdit
-    ? 'Click on the mask to edit it'
-    : 'Click on the map to create a mask';
+    ? intlConfig.formatMessage({
+        id: `c4r.widgets.featureSelection.clickToEdit`
+      })
+    : intlConfig.formatMessage({
+        id: `c4r.widgets.featureSelection.clickToCreate`
+      });
 
   return (
     <StylesWrapper>
@@ -82,14 +124,18 @@ function FeatureSelectionWidgetUI({
         tooltipPlacement={tooltipPlacement}
       />
       <FeatureSelectionUIDropdown
-        selectionModes={selectionModes}
-        editModes={editModes}
+        selectionModes={selectionModeWithLocales}
+        editModes={editModesWithLocales}
         selectedMode={selectedMode}
         onSelectMode={onSelectMode}
         enabled={enabled}
         tooltipPlacement={tooltipPlacement}
-        tooltipText='Select a mode'
-        menuHeaderText='Choose a selection mode'
+        tooltipText={intlConfig.formatMessage({
+          id: `c4r.widgets.featureSelection.selectMode`
+        })}
+        menuHeaderText={intlConfig.formatMessage({
+          id: `c4r.widgets.featureSelection.chooseSelectionMode`
+        })}
         editDisabled={!geometry}
       />
       {!!geometry && (
@@ -97,11 +143,16 @@ function FeatureSelectionWidgetUI({
           features={[geometry]}
           onSelectGeometry={onSelectGeometry}
           onDeleteGeometry={onDeleteGeometry}
-          disabledChipTooltip='Apply mask'
-          chipTooltip='Clear mask'
+          disabledChipTooltip={intlConfig.formatMessage({
+            id: `c4r.widgets.featureSelection.applyMask`
+          })}
+          chipTooltip={intlConfig.formatMessage({
+            id: `c4r.widgets.featureSelection.clearMask`
+          })}
           tooltipPlacement={tooltipPlacement}
           size={size}
           chipLabel={chipLabel}
+          intl={intlConfig}
         />
       )}
     </StylesWrapper>
@@ -132,7 +183,13 @@ FeatureSelectionWidgetUI.propTypes = {
   onSelectGeometry: PropTypes.func,
   tooltipPlacement: PropTypes.string,
   size: PropTypes.oneOf(['small', 'medium']),
-  chipLabel: PropTypes.string
+  chipLabel: PropTypes.string,
+  intl: PropTypes.object
 };
 
 export default FeatureSelectionWidgetUI;
+
+// Aux
+function getModeKeyForLocale(id) {
+  return id.charAt(0).toLowerCase() + id.slice(1);
+}